fix: iReservation had wrong return.

+++ library-reservation.Application/IReservationService.cs
@@ -8,7 +8,7 @@ namespace library_reservation.Application
     public interface IReservationService
     {
         Task<(List<Reservation>, int TotalRecords)> GetPaginatedReservations(PaginationDTO paginationDTO);
-        int GetReservationPrice(List<ReservationItemDTO> items);
+        decimal GetReservationPrice(List<ReservationItemDTO> items);
 
     }
 }
